+++ com.nxp.timedoctor.core.model/src/com/nxp/timedoctor/core/model/lines/ISRSampleLine.java
@@ -155,7 +155,7 @@ public class ISRSampleLine extends SampleLine {
 					n--;
 					int j = st[n];
 					getSample(j).val = i;
-					getSample(i).val = (double) clr[n];
+					getSample(i).val = clr[n];
 					if (getSample(i).type == SampleType.STOP) {
 						setMaxSampleDuration(Math.max(getMaxSampleDuration(),
 								getSample(i).time - getSample(j).time));
